Exclude defaults for both streams & producer config
